Move response types to objects.*
