+++ objects/errors.go
@@ -1,4 +1,4 @@
-package slack
+package objects
 
 import (
 	"encoding/json"
