+++ slack.go
@@ -1,6 +1,7 @@
 //go:generate go run internal/cmd/genevents/genevents.go
 //go:generate go run internal/cmd/genscopes/genscopes.go
 //go:generate go run internal/cmd/genmethods/genmethods.go
+//go:generate go run internal/cmd/genserver/genserver.go
 
 // Package slack implements a REST client for Slack services.
 package slack
